+++ packages/react-internal/src/components/Link/index.ts
@@ -1,0 +1,3 @@
+export * from './Link';
+export * from './Link.base';
+export * from './Link.types';
