--- packages/react-link/src/next/index.ts
@@ -1,3 +1,0 @@
-export * from '../components/Link/index';
-export { Link } from './Link';
-export { LinkBase } from './LinkBase';
